spec: revert change to replace_content for chrome
The specail case for chrome is for the Mac only.  Selenium is able to
send command-a or control-a to chrome just fine on linux. So, tests that
pass on our linux build systems were failing locally on a Mac.

Change-Id: I8733de2615d54b75c4d73e88ba3f54b671c28274
Reviewed-on: https://gerrit.instructure.com/132449
Tested-by: Jenkins
Reviewed-by: Gentry Beckmann <gbeckmann@instructure.com>
Product-Review: Robert Lamb <rlamb@instructure.com>
QA-Review: Robert Lamb <rlamb@instructure.com>