This change makes the default_view setting available to pandarus
While the 'default_view' setting is json api and and is settable via
the api it is not in the documentation completely.  This adds it to the
documentation so that pandarus can read it while generating the gem.

closes GH-1170
closes CORE-845

Change-Id: Ief9ac738d1ea70a3dc076f9e222ca763a50cf9d2
Reviewed-on: https://gerrit.instructure.com/137424
Reviewed-by: Rob Orton <rob@instructure.com>
Tested-by: Jenkins
QA-Review: Matthew Edwards <medwards@instructure.com>
Product-Review: Matthew Edwards <medwards@instructure.com>