fix extended object in i18nObj
This change makes the i18n object independent of underscore,
which was not being imported into the file. The change is covered
by existing specs and is functionally the same before and after
the change. Those specs were indirectly including underscore,
which becomes globally available as a result. This means that
only specs which import this I18n object and NOT underscore would
fail. This does not happen when the suite is run in its entirety.

test plan:
* ensure Jenkins passes

Change-Id: I7d76303e28236d893b66dbed5575e2e355cde738
Reviewed-on: https://gerrit.instructure.com/136018
Tested-by: Jenkins
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Product-Review: Jeremy Neander <jneander@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>