Fix logging and notification searching in admin tools
This was previously doing string comparison of human readable dates to
determine if the search was valid. Fix this so that it will now check
javascript `Date` objects instead, as well as not closing the modal and
actually displaying an error message if you try to perform a search with
the start date ahead of the end date.

This also changes functionality so an error will be raised if a date was
invalid, instead of silently defaulting to no date behind the scenes.

Fixes COMMS-568

Test Plan
  - Send some messages to a user
  - Go to the admin tools page (admin -> `account_name` -> admin tools )
  - Click on the View Notifications tab
  - Input the user you sent the messages to
  - click on the user from the new list of users
  - Input an invalid start and end date (such as `BANANA`), and notice
    that it prints an error for the invalid dates. Note that the error
    will not disappear unless you select the form, ie not if you only
    click the calendar button. This will be a separate ticket.
  - Select a time window with the start time being after the end time,
    and click 'Find'. Notice that an error message pops up indicating
    that this is invalid
  - Select a time window between when you sent the messages, and click
    find again. Notice that the modal will close and the that the
    messages show up.
  - Run the invalid dates with the `logging` tab as well, and make sure
    that those also properly display the error messages when invalid
    dates are selected

Change-Id: Idecafe5bc02f62ca0a5b89a584ad55b0fa59adbe
Reviewed-on: https://gerrit.instructure.com/138004
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Reviewed-by: Venk Natarajan <vnatarajan@instructure.com>
QA-Review: Venk Natarajan <vnatarajan@instructure.com>
Product-Review: Landon Gilbert-Bland <lbland@instructure.com>