Perform first step to refactoring spec methods.
This creates a new course_factory file (course_factory_v2),
and moves the old version to a "_deprecated" file so as to
deter usage.  The hope is this will allow migrating to the new
version on a file-by-file (or even a single-test) basis.

Refs COMMS-881

Test Plan:
* Tests still pass

Change-Id: Ifde2faeaf85a21f65cd0f0129d9b9797de778314
Reviewed-on: https://gerrit.instructure.com/142410
Reviewed-by: Robert Lamb <rlamb@instructure.com>
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>