consolidate gradebook grid source files
This was a part of the endgame for Entering Grades. Consolidating the
grid source files is intended to help further extraction of grid
behavior in the future.

This change consists purely of file relocation and import path updates.
Webpack will have resolved the changes and would complain if this was
not done correctly.

test plan:
 * ensure Jenkins passes
 * super quick smoke test of New Gradebook

Change-Id: I64971c1f3e138b9fb25ace2e371f4c617dd1670d
Reviewed-on: https://gerrit.instructure.com/143060
Tested-by: Jenkins
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>