Fix position handling for pinned discussions.
This makes sure duplication handles positions for pinned
discussions, and we render discussions moved to the pinned
section as the last one in that section.

This should result in the ordering of pinned discussions
not changing just because we hit reload, and keeps the
frontend and backend on the same page on the positioning
of pinned discussions.

Also this refactors the SSD selenium specs since they
were being flaky.  The only thing I can think of is
having 4 discussions on a course page is too many,
so I refactored things to create new courses and
discussions if I needed a different course than was
there already.

Fixes COMMS-954

Test Plan:
* Have section specific discussions enabled.
* Have some pinned discussions.
* Duplicate some.  Note the order of the discussions.
* Reload.  The order should be preserved.
* Make one of these pinned discussions *graded*
* Go to the assignments page, and duplicate the
  corresponding assignment.
* Go back to the discussions page, the new graded
  discussion should be right below the original.
* Mix and match this with some other operations,
  it should all still work, and stuff in the
  pinned section shouldn't mysteriously move
  around just because of a reload.

Change-Id: I5f699689290d5c70158940865e2193ea8f7d403a
Reviewed-on: https://gerrit.instructure.com/143365
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
QA-Review: Aaron Kc Hsu <ahsu@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>