first part of RTL changes
closes: CORE-1171

I split all the  changes to do RTL into 2 commits. This one
contains all of the things that actually change css or javascript.
That way, the test plan on the other, with much more changes to source
code, can be to simply verify that the generated css output is exactly
the same as it was before for all of our existing CSS files, and that
it only changes things for the RTL set of stylesheets.

test plan:
* absolutely nothing should change for anyone that doesn't have
  the RTL feature flag(s) turned on.
* turn on the RTL feature flag.
* these changes should make it so stuff looks better in RTL,
  (you'll need to also checkout the following commit to see *all* of the
  RTL changes)

Change-Id: Ic24ec8dae454709d40f27c2aec80a40ff81c8ef0
Reviewed-on: https://gerrit.instructure.com/143971
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>