Flip right/left instructure-icons in RTL
Closes: CORE-1213

This is just temporary until InstUI and instructure-icons handle arrow
flipping theirselves. Tickets have been created and they are working on
an RFC for how to handle it but until they release a version that does
it automatically for us, this works.

I specifically made a test for it so that when whoever and whoever we do
upgrade either library it will remind us that we need to take this out.

But this will make it so people testing out the new RTL stuff aren’t
thrown off by the arrows pointing the wrong direction

Test plan:
* turn on RTL
* go to the account course/user search page (accounts/x/users)
* look at the pagination at the bottom of the page, the arrows
  should point in the right direction

Change-Id: I46da5647a7e6143e88ea31deb4449029a90a2b8d
Reviewed-on: https://gerrit.instructure.com/143990
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>