Quote student ID in hash passed to SpeedGrader
Wrap the student ID passed to SpeedGrader from the submission detail
dialog (in old Gradebook) and submission tray (in new Gradebook) in
quotes. This prevents it from being treated as an integer and
consequently, in the case of IDs larger than Javascript's max integer
size, garbled (whereupon SpeedGrader would fall back to selecting the
first student in the course instead of the desired student).

fixes GRADE-930

Test plan:

To test this, you will need to create a user with a very large ID, with
a name that ensures they will be sorted after at least one other
student. The easiest option is to do something like the following in the
Rails console:

> user = User.create!(id: 36990000000108595, name: "ZZZ Big Id User")
> ps = Pseudonym.create!(user: user, unique_id: 'big_id_user',
    account: user.account, password: 'big_id_user',
    password_confirmation: 'big_id_user')

(You don't have to stick entirely to the above; the main requirement is
that the ID be greater than 9007199254740991--the value of
Number.MAX_SAFE_INTEGER in Javascript--and that the user be named so as
to be preceded by someone else when sorted.)

Once the user (hereafter "Large-ID-User") has been created:
* Set up a course with at least one student (with a name alphabetically
  preceding the user you just added) and one assignment. Have it use
  old Gradebook for now.
* Add Large-ID-User to the course. I recommend adding them by login
  ID so you can be sure you're adding the user you created and not
  another new user.
* Open Gradebook and view the submission details for Large-ID-User.
* Click on the "More details in the SpeedGrader" link.
* When the SpeedGrader page opens, verify that it is showing the details
  for Large-ID-User (e.g., the proper grade and any comments or
  submissions) and not some other user.
* Switch between users in SpeedGrader using the arrows and pop-up menu
  and verify that it still works as expected.
* Finally, switch the course to use new gradebook and repeat the above
  checks via the SpeedGrader link in the submission tray for
  Large-ID-User.

Change-Id: I17994f4cba86d0d48bb7e0f35f1de8bc43ce21d1
Reviewed-on: https://gerrit.instructure.com/144090
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Tested-by: Jenkins
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>