correct animation for load prior dates button
When clicking on the load prior dates button in the dashboard list view,
the focus should stay on the button and the button should stay in view.
This means the scroll position will jump to the top of the page when the
new items are loaded.

If intead the window is scrolled to load the past by the scroll wheel
or keyboard scrolling buttons, the scrolling position should be
maintained and not jump to the top.

closes ADMIN-781

test plan:
- have lots of items in the past
- make sure loading past items behaves as above

Change-Id: Ib1e38e9eb487da56d395ae7f0ce3cf747defcb21
Reviewed-on: https://gerrit.instructure.com/144390
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>