Allow Assignment Description box to auto-expand to fit contents
Fixes RD-4407

When a teacher views an assignment page, if there is a file linked
in the description, and they click the button to preview with
DocViewer, the preview doesn't all fit into the description div, and
they have to manually enlarge the area to see the preview.  This
change makes the description area size itself to its contents, so
it expands to fit the document preview.  It also increases the
default height of the iframe from 400px to 800px.

Test plan:
- Create an assignment with a file link in the description.
- Click the icon to preview the file.
- Confirm that the DocViewer preview is visible.

Change-Id: I7bfd0ca35a671f2c55f2351ce3909580e0eb258f
Reviewed-on: https://gerrit.instructure.com/144403
Reviewed-by: Daniel Nehring <dnehring@instructure.com>
Tested-by: Jenkins
Reviewed-by: brian kirkby <bkirkby@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Caleb Guanzon <cguanzon@instructure.com>