Handle RTL in inline styles
Closes: CORE-1243

These are all the places that we are using in line styles that w
ould be affected in right to left. This change does absolutely nothing 
to our existing left to right CSS it just flips it in right to left.

Test plan:
Because this change does not change anything for our left-to-right users
We can safely merge it without having to manually check these pages 
Just likewe did With the other automated CSS changes

Change-Id: I62841f4b47925b4d9594c3ca914fa3e5d53fd7b0
Reviewed-on: https://gerrit.instructure.com/145447
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>