planner's initial load fills screen more reliably
Turns out this wasn't specific to ie, and the masquerading border had
some side effects about how tall the document is. So instead of checking
whether the document is taller than the screen, we'll check whether
the fixed element at the bottom of the items list is on screen. This
means we'll try to push the "load more" button below the screen, which
feels natural.

fixes ADMIN-925

test plan:
- have 9 items on today and some future items.
- if the browser height is taller than the 9 items, it should load
the future items until the window is full.
- if the browser height is shorter than the 9 items, it should only
load the 9 items.

Change-Id: Id9cb4c316dc2e96286e9a4e509ff6f8d9b2f5384
Reviewed-on: https://gerrit.instructure.com/149300
Tested-by: Jenkins
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>