spec: Don't automatically insert Authorization header
Instead, we'll only add an auth token if a dummy Authorization header
is already there. Then we'll update that with a valid auth token for
the user specified in the Auth-User-Id header.

If the Authorization header is present, but Auth-User-Id isn't, we
assume user ID 1.

closes QA-514

Change-Id: I02bdcec1bd432e2446236375eeded7853235af7c
Reviewed-on: https://gerrit.instructure.com/155190
Reviewed-by: Robert Lamb <rlamb@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Tested-by: Jenkins
Product-Review: Tucker McKnight <tmcknight@instructure.com>