Use index-based for loop instead of for-in in SG
Using for-in on an empty array (from test data) was pulling in Ember
properties like firstObject for some reason, which was causing a spec
failure when we tried to access those objects in the array. Stop using
for-in so that the test stops failing.

Test plan:
- Specs pass

Change-Id: If9ff1b6785b4e1f113fe0195051696c404e8e75e
Reviewed-on: https://gerrit.instructure.com/155648
Tested-by: Jenkins
Reviewed-by: Gary Mei <gmei@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Indira Pai <ipai@instructure.com>
Product-Review: Adrian Packel <apackel@instructure.com>