spec: add global sandbox
This is an updated form of the magic `this.stub` magic present in the
legacy QUnit specs. Subsequent commits will replace those calls with the
updated form, which is a global `sandbox` object with `stub`, `spy`,
`mock`, and `clock` methods which are automatically verified and
restored at the end of each spec.

Replacing the older methods with the updated methods will allow these
legacy specs to be updated for modern testing.

Change-Id: Id409756ce5f5cb60c52a0735d385e5648fe9dd8f
Reviewed-on: https://gerrit.instructure.com/157014
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Jeremy Neander <jneander@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>