Submit trumps unchecking complete in planner
The user marks an item as complete, then incomplete, then submits the
assignemt. Currently, the planner_override.marked_complete is used to
set the item to incomplete.  The behavior we want if for the submission
to override the student's uncheck, until they uncheck it again.

This is a corner case, probably seen only in a demo
environment.

Update planner_override.marked_complete when the
submission is submitted.  The ugly part is that we have to look at the
submission's assignment for discussion_topic and quiz to find
the right override until assignments are handled more uniformly.

closes ADMIN-1262

test plan:
  - have an assignment
  - have student check as complete, then uncheck as incomplete
  - have student submit the assignment
  - have the student refresh the planner
  > expect planner to show assignment as complete
  - have the student uncheck the assignment as incomplete
  - have the student refresh the planner
  > expect planner to show the assignment as incomplete
  - have the teacher grade the assignment
  - have the student refresh the planner
  > expect the assignment to still be unchecked as incomplete
  - repeat with a graded discussion, a quiz that requires grading,
    and an auto-graded quiz

Change-Id: I59be1e39ef35ecbd8bd60d3075d023154de49609
Reviewed-on: https://gerrit.instructure.com/159125
Tested-by: Jenkins
Reviewed-by: Mysti Sadler <mysti@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Christi Wruck