Remove absolute positioning for parent account creation
The absolute position caused it to overlap on mobile devices
when the keyboard opened up, but only on Android it seemed.
Oh and it did it in responsive testing mode on desktop Safari
and desktop Chrome.  The only place it seemed to not cause
real issues was on the iPhone X simulator.

closes CORE-1639

Test Plan:
  - Set up an account to have self registration for observer
    accounts only.
  - Open up a browser in responsive mode to simulate a mobile
    device.
  - Move the vertical size smaller emulating a keyboard being
    opened.
  - Notice that it requires vertical scrolling to get to the
    parent registration link.

Change-Id: I9e3c0d659d2a918a74d1653109e332007c21026b
Reviewed-on: https://gerrit.instructure.com/159546
Tested-by: Jenkins
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
QA-Review: Rohan Cheeniyil <rcheeniyil@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>