speed up dashcards by only loading tabs shown
Closes: CORE-1781

This removes the need to precacluate permissions for the dashboard card
response by instead just calculating the permissions for the tabs we are
actually going to show (ie: only discussions, assignments,
announcements, and files). Since there is a lot less work to just
calculate those 4, and since we’re only showing a max of 20 courses,
it’ll be faster/easier to maintain to just lazy-compute them.

It also changes the logic in the Course::tabs_available method to not
even do the permission check if we’ve already ruled out that we want to
show this tab, thus preventing a ton of permission lookups that don’t
need to happen.

And lastly, it adds the ability to just check the active_record_types
We care about so we don’t do more ‘count’ SQL queries than we need to.

So the net result is that the request to load the dashcards should be
a lot faster since we only check the permissions we absolutely have to.
and we don’t preload or compute anything that we don’t.

Test plan:
* the XHR for /dashboard-cards when you go to the dashboard page
  Should only include course links to the tabs that are shown on the
  dashcards. (ie: only discussions, assignments, announcements, & files)
* app server/db/redis server load from the dashboard page should go down

Change-Id: I1869b0ca3c3b894a76a9c76ebb74cc4371911753
Reviewed-on: https://gerrit.instructure.com/161982
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins
QA-Review: James Williams  <jamesw@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>